% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{args_to_string}
\alias{args_to_string}
\title{Stringify the list of items}
\usage{
args_to_string(args)
}
\arguments{
\item{args}{Named list of arguments}
}
\value{
Character
}
\description{
Stringify the list of items
}
\examples{
x <- lst(
  a = 1,
  b = 33,
  c = "sfdf"
)
args_to_string(x)
> "a = 1, b = 33, c = sfdf"
}
