% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapvis-class.R
\name{init_mapvis}
\alias{init_mapvis}
\title{Initialise mapvis}
\usage{
init_mapvis(params, vafs)
}
\arguments{
\item{params}{Tbl with mutation_id, and a, b, c params}

\item{vafs}{Tbl with 3 columns: mutation_id, region, VAF (long format)}
}
\value{
mapvis object.
}
\description{
Initialise mapvis
}
